Updated sales report

+++ src/Customer.java
@@ -3,9 +3,9 @@ public class Customer {
 	private long customerId;
 	private String cutomerName;
 	private Boolean isMember;
-	private int contactNumber;
+	private long contactNumber;
 
-	public Customer(long customerId, String customerName, Boolean isMember, int contactNumber) {
+	public Customer(long customerId, String customerName, Boolean isMember, long contactNumber) {
 		this.customerId = customerId;
 		this.cutomerName = customerName;
 		this.isMember = isMember;
@@ -20,7 +20,7 @@ public class Customer {
 		return this.cutomerName;
 	}
 
-	public int getContactNumber(){
+	public long getContactNumber(){
 		return this.contactNumber;
 	}
