+++ src/CustomerMgr.java
@@ -6,7 +6,7 @@ public class CustomerMgr {
 
     }
 
-    public Customer findCustomer(List<Customer> customers, int contactNumber) {
+    public Customer findCustomer(List<Customer> customers, long contactNumber) {
         int custSize = customers.size();
         for (int i = 0; i < custSize; i++) {
             if (customers.get(i).getContactNumber() == contactNumber) {
